Add support for uploading avro schemas to schema registry
Adds support for uploading local avro schemas to a schema registry.

This is useful if you wish to version control avro schemas on git and
have your CI do the registration to schema registry.

Additionally, I've added tests using sbt scripted to test the plugin.
It requires a schema registry running on localhost:8081

Signed-off-by: Ian Duffy <ian.duffy@zalando.ie>